backend/requirements.txt
aiofiles==0.7.0
fastapi==0.68.1
pdf2image==1.16.3
Pillow==9.5.0
PyPDF2==3.0.1
python-multipart==0.0.5
slowapi==0.1.5
uvicorn==0.15.0